requirements.txt
litserve==0.2.11a1
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0 